fix: align schedule date time input model files
