+++ lib/presentation/schedule_create/schedule_date_time/input_models/schedule_time_input_model.dart
@@ -1,17 +1,15 @@
 import 'package:formz/formz.dart';
 
-/// Validation errors for the [ScheduleDateInputModel] [FormzInput].
-enum ScheduleDateValidationError {
-  empty,
-}
+/// Validation errors for the [ScheduleTimeInputModel] [FormzInput].
+enum ScheduleTimeValidationError { empty }
 
-class ScheduleDateInputModel
-    extends FormzInput<DateTime?, ScheduleDateValidationError> {
-  const ScheduleDateInputModel.pure([super.value]) : super.pure();
-  const ScheduleDateInputModel.dirty([super.value]) : super.dirty();
+class ScheduleTimeInputModel
+    extends FormzInput<DateTime?, ScheduleTimeValidationError> {
+  const ScheduleTimeInputModel.pure([super.value]) : super.pure();
+  const ScheduleTimeInputModel.dirty([super.value]) : super.dirty();
 
   @override
-  ScheduleDateValidationError? validator(DateTime? value) {
-    return value == null ? ScheduleDateValidationError.empty : null;
+  ScheduleTimeValidationError? validator(DateTime? value) {
+    return value == null ? ScheduleTimeValidationError.empty : null;
   }
 }
